# docs/openapi.json
{
  "components": {
    "schemas": {
      "CacheEntry": {
        "description": "One row in `cache_entries`.",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "repoId": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "updatedAt": {
            "format": "int64",
            "type": "integer"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "version",
          "scope",
          "repoId",
          "updatedAt",
          "locationId"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Error response body shape — matches `error_response`'s emitted JSON.\nExposed as a typed struct (rather than purely `json!`) so `OpenAPI`\ncan document it; the response builders in this module stay on the\nmacro path so the diff is small.",
        "properties": {
          "message": {
            "description": "Human-readable error message.",
            "type": "string"
          },
          "statusCode": {
            "description": "HTTP status code, repeated in the body for clients that read\nit from JSON (matches upstream h3 `createError`).",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "statusCode",
          "message"
        ],
        "type": "object"
      },
      "ListBody": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StorageLocation"
            },
            "type": "array"
          },
          "itemsPerPage": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "page": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items",
          "page",
          "itemsPerPage"
        ],
        "type": "object"
      },
...
      "StorageLocation": {
        "description": "One row in `storage_locations`.",
        "properties": {
          "folderName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lastDownloadedAt": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "mergeStartedAt": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "mergedAt": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "partCount": {
            "format": "int64",
            "type": "integer"
          },
          "partsDeletedAt": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "folderName",
          "partCount"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "description": "Bearer token equal to the server's `MANAGEMENT_API_KEY`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Operator-facing REST surface for the gha-cache-oxide cache server. All endpoints require an `Authorization: Bearer <MANAGEMENT_API_KEY>` header.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "gha-cache-oxide management API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cache-entries/{id}": {
      "delete": {
        "description": "Removes the cache entry, its backing storage location row, and the folder on the storage adapter.",
        "operationId": "delete",
        "parameters": [
          {
            "description": "Cache entry id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete cache entry",
        "tags": [
          "cache-entries"
        ]
      },
      "get": {
        "description": "Note: the literal-segment route `GET /cache-entries/match` is\nregistered alongside this one. axum (matchit-based) prefers literal\nsegments over `{id}` params, so `/match` always reaches\n`match_endpoint` even when an entry happened to have the literal id\n\"match\". A test pin in `tests/management.rs` guards this routing\ninvariant.",
        "operationId": "get_one",
        "parameters": [
          {
            "description": "Cache entry id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheEntry"
                }
              }
            },
            "description": "The cache entry"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get cache entry",
        "tags": [
          "cache-entries"
        ]
      }
    },
...
    "/storage-locations": {
      "get": {
        "description": "Paginated list of storage_locations rows.",
        "operationId": "list",
        "parameters": [
          {
            "description": "1-based page number (default 1)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Page size (default 100, max 500)",
            "in": "query",
            "name": "itemsPerPage",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListBody"
                }
              }
            },
            "description": "Page of storage locations"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid Bearer token"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "MANAGEMENT_API_KEY env var unset"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List storage locations",
        "tags": [
          "storage-locations"
        ]
      }
    },
    "/storage-locations/{id}": {
      "delete": {
        "description": "Removes the storage_locations row (cascading to any cache_entries pointing at it) and best-effort deletes the underlying blob folder.",
        "operationId": "delete_one",
        "parameters": [
          {
            "description": "Storage location id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete storage location",
        "tags": [
          "storage-locations"
        ]
      },
      "get": {
        "operationId": "get_one",
        "parameters": [
          {
            "description": "Storage location id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageLocation"
                }
              }
            },
            "description": "The storage location"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No row matched"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get storage location",
        "tags": [
          "storage-locations"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Management API root",
      "url": "/management"
    }
  ],
...
}